simplify layout code
